Update action.yml
Add inputs regarding artifact-name and if the action should upload the artifact as input.

Also move the generated plugin in a folder named after the plugins slug, so that the plugin will not end up in a subdirectory with the artifact name.

Also output path to the zipfile directory, ID of the uploaded artifact and URL of the uploaded artifact.